# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KeyStone AD Management API",
        "description": "API for managing on-premise Active Directory (Multi-Domain)",
        "contact": {
            "name": "Keystone Admin",
            "email": "admin@keystone.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user against Active Directory",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Authentication successful"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc..."
                                        },
                                        "user": {
                                            "properties": {
                                                "username": {
                                                    "type": "string",
                                                    "example": "jdoe"
                                                },
                                                "display_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jdoe@ncc.lab"
                                                },
                                                "roles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "L2",
                                                        "Domain Users"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Invalid domain"
                    },
                    "500": {
                        "description": "LDAP connection error"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logs out the current user",
                "operationId": "c4eddaf1a2e990cc90e4326c0d731d41",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/passwords/reset-standard": {
            "post": {
                "tags": [
                    "Password Management"
                ],
                "summary": "Reset a standard user's password",
                "operationId": "9621547ad719b602f5d8df74fd390e88",
                "requestBody": {
                    "description": "Standard password reset data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetStandardPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/passwords/reset-admin": {
            "post": {
                "tags": [
                    "Password Management"
                ],
                "summary": "Reset a privileged admin account's password",
                "operationId": "56650a0abe7978b99091021fb51f3d3f",
                "requestBody": {
                    "description": "Admin password reset data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetAdminPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin password reset successfully"
                    },
                    "404": {
                        "description": "Admin user not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "title": "Login Request",
                "required": [
                    "username",
                    "password",
                    "domain"
                ],
                "properties": {
                    "username": {
                        "description": "The user's Active Directory username.",
                        "type": "string",
                        "example": "jdoe"
                    },
                    "password": {
                        "description": "The user's Active Directory password.",
                        "type": "string",
                        "format": "password",
                        "example": "P@ssword123!"
                    },
                    "domain": {
                        "description": "The Active Directory domain to authenticate against.",
                        "type": "string",
                        "example": "ncc.local"
                    }
                },
                "type": "object"
            },
            "ResetAdminPasswordRequest": {
                "title": "Reset Admin Password Request",
                "required": [
                    "domain",
                    "baseSamAccountName"
                ],
                "properties": {
                    "domain": {
                        "description": "Domain of the user (e.g., ncc.lab)",
                        "type": "string"
                    },
                    "baseSamAccountName": {
                        "description": "The sAMAccountName of the *standard* user (e.g., '118155', not '118155-a')",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ResetStandardPasswordRequest": {
                "title": "Reset Standard Password Request",
                "required": [
                    "domain",
                    "samAccountName"
                ],
                "properties": {
                    "domain": {
                        "description": "Domain of the user (e.g., ncc.lab)",
                        "type": "string"
                    },
                    "samAccountName": {
                        "description": "sAMAccountName of the standard user to reset",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
...
}